+++ diesel_cli/tests/migration_run.rs
@@ -44,7 +44,7 @@ fn migration_run_inserts_run_on_timestamps() {
                        "DROP TABLE users");
 
     let migrations_done: bool = select(sql::<Bool>(
-            "EXISTS (SELECT * FROM __diesel_schema_migrations)"))
+            "EXISTS (SELECT * FROM __diesel_schema_migrations WHERE version >= '1')"))
         .get_result(&db.conn())
         .unwrap();
     assert!(!migrations_done, "Migrations table should be empty");